Create an error message for when the repo is first cloned and ./_build.sh is run.  In particular, the user is required to create a _pagespeed directory or make a symlink pointing elsewhere.  Also, since this is no longer hosted on github there isn't a need anymore for static redirect pages and the alias submodule can be used since we generate the alias pages and sync to app engine.

+++ _pagespeed.sh
@@ -11,6 +11,15 @@ output_dir=_pagespeed
 
 cd $(dirname ${BASH_SOURCE[0]})
 
+if [ ! -d ${output_dir} ]; then
+  echo
+  echo "Output directory ${output_dir} doesn't exist."
+  echo
+  echo "Create one or use a symlink to point elsewhere..."
+  echo "for example to a directory served by a web server."
+  exit 1
+fi
+
 rm -rf ${output_dir}/*
 
 for dir in $(find ${input_dir} -type d); do
